+++ user.go
@@ -2,5 +2,5 @@ package pythia
 
 type User struct {
 	Salt, DeblindedPassword []byte
-	version                 int64
+	version                 uint
 }
